Move plotting functions to plotting.py
