+++ rlpy/tools/__init__.py
@@ -1,3 +1,4 @@
 from .general_tools import *
-from .priority_queue_with_novelty import PriorityQueueWithNovelty
 from .general_tools import __rlpy_location__
+from .plotting import *
+from .priority_queue_with_novelty import PriorityQueueWithNovelty
